Add LogisticRegressionPipeline - Change which dataset is selected for tuning and compare stages. - Modify README.

+++ src/main/scala/upm/bd/transformers/Preprocesser.scala
@@ -66,7 +66,7 @@ class Preprocesser(delayThreshold: Int = 15) {
     MyLogger.info("Adding 'OverDelay' and 'WeekEnd' columns")
     df = df.select(
       $"*",
-      ($"ArrDelay" > lit(delayThreshold)).as("OverDelay"),
+      ($"ArrDelay" > lit(delayThreshold)).as("OverDelay").cast("int"),
       (col("DayOfWeek") === 6 || col("DayOfWeek") === 7).as("WeekEnd")
     )
